F1: change typedef of uint32 to unsigned long int

+++ STM32F1/cores/maple/libmaple/libmaple_types.h
@@ -42,7 +42,7 @@ extern "C" {
 
 typedef unsigned char uint8;
 typedef unsigned short uint16;
-typedef unsigned int uint32;
+typedef unsigned long int uint32;
 typedef unsigned long long uint64;
 
 typedef signed char int8;
